+++ test/tiles/beddb_test.py
@@ -1,5 +1,6 @@
-import clodius.tiles.beddb as hgbe
 import os.path as op
+
+import clodius.tiles.beddb as hgbe
 
 
 def test_get_tiles():
@@ -23,3 +24,11 @@ def test_name_in_tile():
     tiles = hgbe.tiles(filename, ["x.1.0", "x.1.1"])
 
     assert "name" in tiles[0][1][0]
+
+
+def test_tileset_info():
+    filename = op.join("data", "geneAnnotationsExonUnions.1000.bed.v3.beddb")
+
+    tsinfo = hgbe.tileset_info(filename)
+
+    assert "chromsizes" in tsinfo